virt-launcher: Fix host-device metadata key calculation
In order to calculate the proper key for host-devices, their fixed
alias-prefix needs to be removed.

The removal was buggy in case the same alias-prefix was used in the
original name of the device.

To resolve the issue, the alias-prefix of the relevant host-device
is removed only from the name prefix (once).

fixes: https://github.com/kubevirt/kubevirt/issues/10401

Signed-off-by: Edward Haas <edwardh@redhat.com>