rbac: Add instancetype.kubevirt.io:view ClusterRole
This ClusterRole provides read only access to the cluster-scoped
VirtualMachineCluster{Instancetype,Preference} resources within a
cluster.

This is required as the existing aggregated kubevirt:view ClusterRole
does not provide access to cluster-scoped resources when bound to a user
using a RoleBinding. The tests are extended to verify this.

As such we need a seperate ClusterRole to provide access to our cluster
scoped resources that can be bound with a ClusterRoleBinding without
providing access to all of the resources listed in the aggregated view
ClusterRole.

Co-authored-by: Felix Matouschek <fmatouschek@redhat.com>
Signed-off-by: Lee Yarwood <lyarwood@redhat.com>