Set memory limits if there is a resource quota with limits.
When a ResourceQuota with memory limits set is associated to
a namespace, if vm creators try to start a vm, the vmi remains
is Pending state, due to missing memory limits.
In fact, a resource quota with memory limits requires every pod
in that namespace to have memory limits set.
Currently, KubeVirt is not setting that limits for every vmi, but
only for the vmi with dedicated CPU.

The virt-launcher overhead is hard to calculate, and it can vary
depending on the vmi spec.
Several testing are ongoing to extract some values, but they require
a lot of time.
This commit introduces an initial easy implementation to avoid the
Pending state of the vmi.
This initial implementation consists in setting limits to the double
of the memory requests.
This will only happen if all following are true:
- the vm does not have any memory limits set.
- the `AutoResourceLimits` FG is enabled.
- there is a resource quota with memory limits associated to the namespace in which the vm is created.

Cluster admins can override the `2x` behaviour, adding the label
`alpha.kubevirt.io/auto-memory-limits-ratio: <custom_float_value>` to the namespace.

Signed-off-by: fossedihelm <ffossemo@redhat.com>