Set cpu limits if there is a resource quota with limits.
Currently, there is a mechanism by whereby if there is a Resource Quota with memory
limits and the `AutoResourceLimits` FG is enabled, Kubevirt will automatically
set the memory limits in the virt-launcher pod.

With CPU limits, another behavior has been implemented:
define a namespace label selector(autoCPULimitNamespaceLabelSelector) that instructs in which namespace the automatic cpu
limits will be set.

The same functionality of the memory limits can be applied to the cpu limits too, so
that if the AutoResourceLimits` FG is enabled, Kubevirt will automatically
set both memory and cpu limits in the virt-launcher pod if a RQ with limits is associated
to the creation namespace.

`autoCPULimitNamespaceLabelSelector` struct will extend the functionality if the admin decides
to enable this on namespaces where RQ is not present.

Signed-off-by: fossedihelm <ffossemo@redhat.com>