net, passt: Drop the attempt to redirect passt logs to virt-launcher log
The current implementation only takes the initial content of the file,
which may or may not be enough for troubleshooting.

As there is no better solution at the moment, this portion of code is
removed.

As a follow up, the user-guide will instruct how to fetch the relevant
logs if needed (from the pod) and possibly read and dump them to the
virt-launcher log on virt-launcher termination.

Signed-off-by: Edward Haas <edwardh@redhat.com>