instancetype: Introduce InferFromVolumeFailurePolicy
By setting this field in the {Instancetype,Preference}Matchers to
'IgnoreInferFromVolumeFailure', it is possible to ignore failures during
inference of an instancetype and/or preference, e.g. because of an
unsupported volume type. Instead of failing the VM creation, the VM is
created and the corresponding matcher is cleared.

This allows to offer virtctl inference of instancetypes and preferences
by default, without confusing users by possibly failing to create VMs
when a valid `virtctl create vm` command was executed.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>