virt-launcher: Dump passt log file on exit instead of domain start
Dump passt logs when virt-launcher app exists if exist instead of
when the domain starts.

Passt log file path is defined in virt-launcher-monitor.go file as
well instead of using the one in package network/domainspec due to
the RAM overhead this package adds to the monitor process (about 5MB).

Signed-off-by: Or Mergi <ormergi@redhat.com>