e2e: Fix flaky leftover `err` variable
The `err` variable is not valued in the "they should be applied to the launcher pod"
test, but there is an expectation regarding it.
Since `err` is global var and never cleaned, this means that we are checking the err value of the previous test.
If the previous test is a negative test this can cause a failure, even if there is no failure.
i.e. if the previous test is "[test_id:3113]should failed to the VMI creation", which
expect that the creation of a vmi fail if there are no memory request, there will be a failure.

Signed-off-by: fossedihelm <ffossemo@redhat.com>