virt-controller, template: Add missing initializers in tests
The template unit tests have currently a leak, where one test
initialization may be used in a follow up test in case the relevant
variables have not been overridden by a new initialization.

This is a local fix to the problem, in the tests that have been
identified as not having a proper initialization.
A more thorough solution should be presented by using a fixture that
assures the variable is overridden.

Signed-off-by: Edward Haas <edwardh@redhat.com>