virt-controller, template: Allow sidecar creation registration
The logic of how a sidecar is to be created and when to create it in the
rendering process are strongly coupled.

In order to allow future sidecar creations and at the same time not
frequently change the flow code, a separation is suggested.
Such a separation also assures the template rendering package is not
polluted by dependencies from all over the project.

This change is limited to the infrastructure that allows new sidecar
creation logic to be added. Following changes will extract some
sidecar creation logic to their relevant packages.

Signed-off-by: Edward Haas <edwardh@redhat.com>