net, setup: Replace `configstate` for the plug/hotplug flow
The `configstate` flow has been minimized with the latest changes in the
code, therefore, to reduce even further the complexity, move its
plug/hotplug flow logic to `netpod`.

A new `state` has been added to `netconf` & `netpod` to manage the pod
configuration state changes. This `state` object is replacing the
`configstate` one and in a following change will allow it to be dropped
completely (when the unplug flow will also flow through `netpod`).

Signed-off-by: Edward Haas <edwardh@redhat.com>