Aggregate DVs conditions on VMI (and so VM)
Today we have really detailed conditions
on DataVolumes but they are not directly reported on
VMs and VMIs.
Conditions provide a standard mechanism for higher-level
status reporting from a controller:
when a VM/VMI is not in the expect state the user is
supposed to check its conditions to understand what's wrong.
Let's aggregate DataVolumes conditions in a top level one
("DataVolumesReady") to simple expose to the user
an aggregated top-level condition on VM/VMI summarizing
more detailed conditions from datavolumes.

This is simplified version consuming only
cdiv1.DataVolumeReady and ignoring DataVolumeRunning
and DataVolumeBound.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>