virt-launcher,converter: Remove redundent check /dev/vhost-net presense
In case a VM requires virtio network, when at least one interface model
is set to virtio explicitly or empty (implicitly), virt-controller template
code will add mount /dev/vhost-net to virt-launcher pod.

The virt-launcher converter code (VM spec to domain XML) checks if /dev/vhost-net
exist, if not and Kubevirt is not set to allow software emulation, it will fail
syncVMI routine and cause re-enqueues.

In Kubernetes, if a pod spec specifies a resource that doesn't exist the pod hang
on scheduling phase and wont start.
In our case, when '/dev/vhost-net' is not present in any node virt-launcher pod wont
start.

Failing in case a VM has virito interface(s) but '/dev/vhost-net' is not present is
redundant, because when virt-launcher converter code runs it means the pod has
started and '/dev/vhost-net' exist.
Also, virt-handler claims ownership of `/dev/vhost-net` [1] just before network setup starts.
If `/dev/vhost-net` is not present, the network setup code will fail and cause the
virt-launcher pod to fail.

Signed-off-by: Or Mergi <ormergi@redhat.com>