instancetypes: Clear VM guest memory when ignoring inference failures
With this change the guest memory of VMs is cleared when inference of an
instancetype was successful and the IgnoreInferFromVolumeFailure policy
was set in the instancetype matcher. This allows to enable inference of
instancetypes in virtctl by default without breaking the previous
behavior by allowing virtctl to provide a fallback amount of memory to
be used when inference fails.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>