multiple bridge interfaces hotplug bug fix
On bridge binding the pod interface name is renamed.
The code allows to setup the same interface multiple times (for example we
can try to hotplug an interface that is already hotplugged, but the VMI
status is not updated yet).
In bridge binding setup we need to take care of the scenario the network
is already configured on the pod and the pod interface name was already
changed.

Signed-off-by: Alona Paz <alkaplan@redhat.com>