virt-controller,tests: Add factory for pod with multus annotations
Generating multus annotations now depends on cluster config.
Passing the tests cluster config to NewPodForVirtualMachine
makes it cumbersome and requires changing all the tests that use it.

Thus, introduce new factory for creating pod with multus annotations.

The tests fake cluster config is moved to global scope to enable access
for tests the needs it; tests that generates multus annotations.

Signed-off-by: Or Mergi <ormergi@redhat.com>