Revert to 100 VMs on the performance lane
PR #10110 introduced an additional container
in each virt-launcher pod to stream out the
logs from the serial console of the guest OS.
This additional container is running a custom
tail binary named virt-tail and so it also
requires resources.
The increased amount of requested resources
(35M of memory for each virt-launcher pod)
was making the performance test lane failing
and so the number of VMs involved in the test was
squezed from 100 to 90 but due to that
the results are now not directly comparable.
Let's revert back to 100.

This requires https://github.com/kubevirt/project-infra/pull/3023
first getting more memory for the -sig-performance
test lanes.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>