virt-config: Add CPUManagerPolicyBetaOptions feature-gate
In k8s 1.22+, when setting CPU Manager Policy: full-pcpus-only.
With the option enabled, the pod will be admitted by the kubelet only
if the CPU request of all its containers can be fulfilled by
allocating full physical cores [0].

In order to support this functionality, adding new
CPUManagerPolicyBetaOptions feature gate.
When enabled - it allows the VM to change the CPU assignment to align
with the Policy chosen. Currently only full-pcpus-only is supported

[0]
https://kubernetes.io/docs/tasks/administer-cluster/cpu-management-policies/#static-policy-options

Signed-off-by: Ram Lavi <ralavi@redhat.com>