virt-controller: Add full-pcpus-only support to IsolateEmulatorThread
When the VMI is created on a node where the CPUManager policy is
full-pcpus-only [0], the virt-launcher pod can only have full physical
cores, otherwise the kubelet will reject it [1].
This can cause internal conflict when setting "IsolateEmulatorThread:
true" on the VMI spec. The reason is that IsolateEmulatorThread adds an
extra emulator CPU to the virt-launcher pod, and can thus cause an
alignment error.

The conditions needed in order to get this alignment error:
* The node has hyperthreading (SMT) enabled.
* Kubelet’s CPUManager is set to static - full-pcpus-only.
* Kubevirt CR's defaultRuntimeClass is set to a high performance
runtimeClass.
* VM is set with dedicatedCPUPlacement and IsolateEmulatorThread
enabled.
* VM requests an even number of CPUs.

In order to avoid this conflict - adding an annotation
"cpuManagerPolicyBetaOptions: full-pcpus-only".
When annotation set and IsolateEmulatorThread enabled, The
virt-controller makes sure that the number of CPUs the virt-launcher pod
requests are even, ensuring that there is no SMT misalignment.
The meaning of this is that instead of IsolateEmulatorThread adding an
extra CPU to the virt-launcher pod, it will add two CPUs when needed in
order to get a total number of even CPUs.

Fixes BZ#2228103

[0]
https://kubernetes.io/docs/tasks/administer-cluster/cpu-management-policies/#static-policy-options
[1] https://github.com/kubernetes/kubernetes/pull/101432

Signed-off-by: Ram Lavi <ralavi@redhat.com>