net, netconf: Process filtered networks
When the network setup is called in the virt-handler reconcile loop, the
networks have been already filtered per the origin of the call (e.g.
migration, initial setup or a hotplug).

Preserve the networks filtered when creating the `netpod` object so
other networks will not be processed.
(e.g. hotplug networks which their pod interface is not ready, should
not be processed)

Signed-off-by: Edward Haas <edwardh@redhat.com>