test reporter: Fix error when passt log not present
The tests reporter raises errors regarding passt log recording failures:
```
19:49:03:   failed to execute command [[ -f /var/run/kubevirt/passt.log ] && cat /var/run/kubevirt/passt.log] on virt-launcher-testvmi-m2mf5-f5vz6, stdout: , stderr: [: missing ']'
19:49:03:   , error: command terminated with exit code 2
```

Since passt log recording command includes terms `[ .. ] && .. ) ` and
not running in a shell, the command is not interpreted correctly and fails.

Fix the the reporter passt logging by executing commands in a shell.

To avoid creating empty log files for non passt VMs, the reporter will record
the command in file if the command stdout is empty.

Signed-off-by: Or Mergi <ormergi@redhat.com>