e2e, nonroot: Execute test
Currently, the `[sig-compute]NonRoot feature` test
is skipped in the CI.

Since kubevirt/project-infra#2418, the CI sets only the
`FEATURE_GATES` environment variable,
and does not set `KUBEVIRT_NONROOT`.

The test is skipped because of a condition defined at
`automation/test.sh` is not met since the above PR.

Execute the test if the Root feature gate is disabled.

Change an assertion in the test, to expect the Root
feature gate to be disabled.

Remove the NonRoot decorator.

Signed-off-by: Orel Misan <omisan@redhat.com>