hack: Deploy CNI plugin artifacts to the cluster nodes
As part of the cluster-sync flow, deploy relevant artifacts to the
nodes.

For now, only a single binary is deployed: the passt binding plugin.

The network binding plugin is conditionally uploaded to the nodes when
`KUBEVIRT_DEPLOY_NET_BINDING_CNI` is set to `true`.
On CI, targets which include `sig-network` in their names are setting
the mentioned env variable and will deploy the CNI plugin.

Signed-off-by: Edward Haas <edwardh@redhat.com>