tests: Introduce a net binding plugin test and `libkvconfig`
This change introduces the passt network binding plugin e2e test.
It is aimed to test the network binding plugin infrastructure, which
utilizes a sidecar and a CNI plugin.

In order to implement the test, the relevant feature-gate
(`NetworkBindingPlugins`) is activated and the network binding is
registered (`passt`).

To help execute these changes on the kubevirt CR, a new `libkvconfig` is
introduced.
The new package is performing the changes using JSON patch

This change also introduces a new ginkgo label `netCustomBindingPlugins`
that should allow control on the test execution.
Only clusters that deploy the network binding plugin CNI should run the
test.

Signed-off-by: Edward Haas <edwardh@redhat.com>