e2e: Remove NonRoot feature test
Currently, the `[sig-compute]NonRoot feature` test
is skipped in the CI.

Since kubevirt/project-infra#2418, the CI sets only the
`FEATURE_GATES` environment variable,
and does not set `KUBEVIRT_NONROOT`.

The test is skipped because of a condition defined at
`automation/test.sh` is not met since the above PR.

Remove the condition from `automation/test.sh`.

Remove the test since it is no longer relevant,
as the Non root implementation is the default.

Remove the `NonRoot` decorator as it is no longer used.

Signed-off-by: Orel Misan <omisan@redhat.com>