e2e, fix multus bridge tests: dummy iface holds the original pod iface MAC
The tests are verifying that when bridge binding is used,
the original pod interface MAC is removed from the pod network namespace.
However, since now the dummy interface holds the MAC,
the test was changed to a more specific verification.
It verifies the MAC is not on the pod interface.

Signed-off-by: Alona Paz <alkaplan@redhat.com>