Fix imageID regex for CRI-O >= 1.28
CRI-O is using a different pattern
for ImageRef (see: https://github.com/cri-o/cri-o/pull/7149 )
now named ImageID (see: https://github.com/cri-o/cri-o/pull/7393 )
and this is breaking a regex we are using to extract
the ImageID from containerStatuses.
Tha is breaking migrations and so the capability
to upgrade virt-launcher on upgrades.

Fixing unit tests.

Fixes: https://github.com/kubevirt/kubevirt/issues/10616

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>