e2e: enable drain migration tests for several cluster topology
Currently, migrations during node drain tests are not available in those
tests where all the nodes are both control-plane and worker; this is because
on the BeforeEach function, all the control-plane nodes are set as unschedulable,
and hence all the nodes are set as unschedulable, resulting in the impossibility
to start a vmi. This trick was done to prevent the vmi to be scheduled initially
on the control-plane node.
We want to avoid this situation in which the vmi will be scheduled on the control-plane
node in topology where there is only one control-plane, because the subsequent drain
will be disruptive.
To prevent the above situation, we will use the PreferredDuringSchedulingIgnoredDuringExecution
term to avoid the vmi to be scheduled in the only one control-plane node;
if there are more than one, even if the vmi will be scheduled there, we can drain it, since
there is at least one control-plane for backup.

Also, moved the `temporaryNodeDrain` function inside the context where it is only used.

Signed-off-by: fossedihelm <ffossemo@redhat.com>