release v1.1.0 perf-scale-data
This commit updates the file name to perf-scale-benchmarks.md
where KubeVirt should publish the perf-scale graphs. The graphs
are modified to have timeline of the releases and change in
kubernetes providers. This will give users idea of how perf-scale
data in changing relative to the KubeVirt code as
well as k8s provider in CI clusters. The perf-scale-benchmarks.md
document is updated to reflect the latest changes.

Signed-off-by: Alay Patel <alayp@nvidia.com>