be less verbose during virt-launcher-monitor startup
virt-launcher-monitor is (eventually) preparing
a file to signal the request for a quick
shutdown to virt-tail if all the others
IPC mechanisms are failing.
This file is in the private dir that is
asyncronously initialized virt-launcher
so some file creation attempts are expected
to fail on virt-launcher-monitor side.
Be less verbose on them to avoid polluting
logs with false error messages.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>