CHANGELOG: Import existing release notes
These are mostly taken from the corresponding git tag, but in a
few cases the information was only present on the GitHub releases
page.

Minimal cleanups were performed: specifically, titles were added
and information such as the list of contributors and the test
results, which didn't render properly as Markdown, were turned
into preformatted text blocks.

As an exception, the release notes for v1.1.0 have been converted
to the format that is going to be used going forward.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>