Restore default hotplug attachment pod request to original values
As a quick fix we modified the requests and limit ratio to
be 1. However this is not ideal since it reserves a lot of
cpu and memory for that container that doesn't really do
anything at all. Its purpose is to attach a volume to the node.

We added the ability to set the requests and limits for all the
supporting containers in (#9422). At this point anyone can set
the ratio to what they want. So lets restore the defaults to the
original values, so that people that don't care don't end up using
a lot of memory for no reason.

Signed-off-by: Alexander Wels <awels@redhat.com>