network-binding: pass the binding config via virtualMachineOptions
The commit extends the options passed from the virt-handler to the
virt-launcher in their grpc communication.
The options will include a mapping between the VMI interface name and
the domain attachment it should use.
VMI interfaces with tap domain attachment will use the tap generator to
generate the libvirt domain.


Signed-off-by: Alona Paz <alkaplan@redhat.com>