Remove Pending flag from test to find leading virt-controller
This test was marked as pending and flaky over three years ago[1] - from
a number of  local runs the test seems to be stable enough. Removing the
Pending flag from the test.

If it flakes again this should be removed as it was not missed for the
last number of years.

[1] https://github.com/kubevirt/kubevirt/commit/b01540f71f2c08857ad28c221603c076fa44000e

Signed-off-by: Brian Carey <bcarey@redhat.com>