virt-controller, vmi, test: Fix unit test
The unit test which has been verifying the scenario where the multus
annotation update is failing due to API unavailability, is incorrect.
The network configuration is not adding the interfaces (only the
networks). Therefore, no pod annotation is really changed.

The test passed because the implementation would have patched the multus
annotation even if there was no need to (e.g. it has not changed).

Signed-off-by: Edward Haas <edwardh@redhat.com>