virt-controller, vmi: Update the multus annotation only if it changed
If the calculated multus annotation is identical to the existing, there
is no need to execute a patch on the VMI object.

This change also resolves a bug that occurs when unplugging the last
secondary network.
Before this fix, when the last network got unplugged, no secondary
network got left, therefore there was no multus annotation generated.
And this led to not patching the VMI object, although it should have.

Signed-off-by: Edward Haas <edwardh@redhat.com>