cpu/mem hotplug: do not send syncVMI in source virt-handler
CPU and memory hotplug actions lead to changes in the VM spec,
which immediately propogated to the VMI object. On virt-handler side
the syncVMI will be called since the VMI spec has changed.
The problem is that SyncVMI will carry the updated CPU/memory topology.
On the live-migration source virt-launcher it can cause failure because
of the inconsistency with the virt-launcher cuurent cpuset

Signed-off-by: Igor Bezukh <ibezukh@redhat.com>