tap binding plugin: Change ownership on tap domainAttachment interfaces
The old code changed the ownership on the macvtap pod interface.
To allow using macvtap binding as a plugin the code is changed to support
any binding that is using tap domain attachment.
If a device exists, the ownership will be changed (masquerade and bridge
are exluded since we know in advance there is no tap device on those
bindings).

Signed-off-by: Alona Paz <alkaplan@redhat.com>