set current cpu topology in VMI status early on VMI creation
CPU current topology was set by the VM controller when
it the VM was started. The bug in this approach is that
when you create the VMI directly, the CPU topology is missing
in the VMI status because the VM controller is not involved.
It also causes a bug in CPU hotplug.

The suggested change sets the CPU topology early in the VMI
creating at the mutating webhook stage.

Signed-off-by: Igor Bezukh <ibezukh@redhat.com>