[test] dump serial console log for debug purposes
Functional test
"it should not skip any log line even trying to
flood the serial console for QOSGuaranteed VMs"
is generating many sequential log lines over the serial
console fetching the output from k8s log api
and ensuring that all the expected logs lines are
there and in the correct order.

Looking at periodic jobs results the test
is sometimes (~10%) failing due to duplicated
chars in the logs, something like:
```
logline 0006195 0006195 0006195 0006195 0006195 0006195 0006195 0006195 0006195 0006195 0006195 0006195 0006195 0006195 0006195
loglline 0006196 0006196 0006196 0006196 0006196 0006196 0006196 0006196 0006196 0006196 0006196 0006196 0006196 0006196 0006196
logline 0006197 0006197 0006197 0006197 0006197 0006197 0006197 0006197 0006197 0006197 0006197 0006197 0006197 0006197 0006197
```

In case of a failure,
let's collect the console log from virt-launcher
pod with an out-of-band mechanism
to be able to understand if the unexpected chars
was already there in the libvirt generated files.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>