launcher, source: Log migration info by default
Currently, the migration info of the source pod
is logged from log level 4.
This information is helpful when debugging problematic
migrations, and is not logged by default.

Log this information with log level to 2, in order for it
to be printed by default.

The log is printed every four seconds.

Signed-off-by: Orel Misan <omisan@redhat.com>