virt-launcher, vcpu: Fix EmulatorThreadPin assign strategy
Currently, when CPUManagerPolicyBetaOption annotation is set to
full-pcpus-only on the VMI, the virt-launcher pod is assigned with extra
two CPU threads meant for emulator-threads [0] - but only if the amount
of CPUs originally requested was even.
This is not aligned with the EmulatorThreadPin assignment on
virt-launcher, where 2 threads are requested regardless of the amount of
vCPUs.

Fix the EmulatorThreadPin assignment to request the second emulator
thread only if the amount of vCPUs is even.

[0] https://github.com/kubevirt/kubevirt/pull/10593

Signed-off-by: Ram Lavi <ralavi@redhat.com>