vmclone: Reject target with same name as source
Currently, it is possible to define the following
VirtualMachineClone spec:
- source.kind is `VirtualMachine`
- target.kind is `VirtualMachine`
- target.name is equal to source.name

In this scenario, the target VM could not be created and
the clone process is stuck.

Add a condition to the validating webhook to block
this scenario.

The following error message will be produced:
```
The request is invalid: spec.target.name: Target name cannot be equal to source name when both are VirtualMachines
```

Signed-off-by: Orel Misan <omisan@redhat.com>