feature-gate: Change CPUManagerPolicyBetaOptions name
Currently the feature gate name CPUManagerPolicyBetaOptions is trying to
align to the same behavior as the CPU-manager policy configured.
Kubevirt should not couple itself with a beta API if not necessary,
cause it could break if CPU-Mangager ever deprecates this API.

Changing the name of the feature gate to `AlignCPUs` in order to focus
on the behavior kubevirt want to achieve, regardless on the
CPUManagerPolicyBetaOptions CPU-Manager policy used on the cluster.

Signed-off-by: Ram Lavi <ralavi@redhat.com>