Change CPUManagerPolicyBetaOptionsAnnotation annotation name
Currently the CPUManagerPolicyBetaOptions Annotation is named after the
CPU-Manager policy it tries to follow.
Kubevirt should not couple itself with a beta API if not necessary,
cause it could break if CPU-Manager ever deprecates this API.

Renaming the Annotation to
`alpha.kubvirt.io/EmulatorThreadCompleteToEvenParity` in order to
focus on the behavior kubevirt wants to achieve, regardless on the
CPUManagerPolicyBetaOptions used on the cluster.

Signed-off-by: Ram Lavi <ralavi@redhat.com>