Remove year from generated code copyright
Problem: the code generation (i.e. deepcopy-gen) regenerates the code
using hack/boilerplate/boilerplate.go.txt as a blueprint for the
copyright header. This file has a placeholder `YEAR` that is replaced by
code generation with the current year. Every beginning of the year this
breaks all PRs since the make generate will complain that the generated
code differs, since the YEAR has changed.

Proposal: remove the YEAR from the template to fix the problem from now
on

Signed-off-by: Daniel Hiller <dhiller@redhat.com>