vmi-create-admitter: Remove CPUManagerPolicyBetaOptions check
Currently the VMI is blocked if the
`kubevirt.io/CPUManagerPolicyBetaOptions` annotation set on the VMI and
the
CPUManagerPolicyBetaOptions feature-gate is not enabled.
Removing this logic from the webhook in order to align with the design
proposal [0]
The meaning of this change is that if the annotation is set on the VMI
manually - then the CPUManagerPolicyBetaOptions:full-pcpu-only behavior
will take place in spite of the feature gate being disabled.

[0] https://github.com/kubevirt/community/pull/247

Signed-off-by: Ram Lavi <ralavi@redhat.com>