fix(vmclone): avoid override variables with returning obj
Create function returns an obj and an error.
In case of error the returned obj is the empty object of the expected type.
For example, the Create function for the Restore resource, in case of
error, will return the empty VirtualMachineRestore obj.

In the clone process, when creating the snapshot and the restore, in
case of `already exists` error we set, respectively, the snapshot name
and the restore name in the clone status.
Since we are overriding the same variable with the returned value,
the snapshot name we are trying to save is an empty string.

This misbehavior can be fixed by defining a new variable for the
returned obj.

Signed-off-by: fossedihelm <ffossemo@redhat.com>
(cherry picked from commit cacb7b30e6a133d9f8ce5ddae2027bc7c449fc39)