Add a notice to virtctl console about logs
Guest serial console output can now be streamed
to cluster logs. This could be eventually enabled
at cluster level being not so visible to VM owners.
Those logs include the boot logs but also the output
of an interactive usage of the first serial console.
No access passwords for the console nor
properly handled passwords (not echoed to the screen)
are going to be collected n the logs.
However, if the VM user types a password by mistake
instead of a command name or if the password is visible
on the screen as a clear text parameter or in the
output of any command, it could be streamed
to cluster logs along with all other screen visible text.

Let's have `virtctl console` emitting a notice
(if LogSerialConsole is not explicitly disabled) for
the user suggesting to use SSH when handling sensitive
informations.

Adding a test covering `virtctl console` output.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>