VMClone: Emit an event in case restore creation fails
When a user creates a VMClone and the target VM already exists,
the clone controller tries to create the `VirtualMachineRestore`,
but is rejected by the restore validating webhook due to the fact
that the target VM already exist.

The user who created the clone has no indication of what is going
wrong.

There is information available about the problem in the logs of
the clone controller, but we assume that it is not accessible to
the user creating the VMClone.

Emit an event in case there was a failure to create the
`VirtualMachineRestore` so it will be reflected on the
VMClone object:

```bash
kubectl describe vmclone <name>
```

or

```bash
kubectl get events
```

Signed-off-by: Orel Misan <omisan@redhat.com>