SRIOV - syncVMI tries to plug an SRIOV interface that is not in the pod
In case SRIOV interface/s exist, syncVMI tries to plug them to the VM.
If it fails, the whole syncVMI fails.
However, there are use cases we konw in advance it is fine to have a
SRIOV interface but not yet manage to plug it.
For example, when hotplugging a SRIOV interface. Until a migration is
done the interface/device is not plugged to the pod, therefore,
syncVM` will fail.
The commit allows ignoring SRIOV interface which are not yet plugged to
the pod and avoiding syncVMI failure.

Signed-off-by: Alona Paz <alkaplan@redhat.com>