libvmi: avoid a boolean argument
A casual reader of

   libvmi.NewFedora(libvmi.WithCloudInitNoCloudUserData(userdata, false))

would find it hard to understand what is "false" about this function
call. Reading deeper would reveal that it means that the
userdata argument is not base64-encoded.

As is commonly the case with functions that have a boolean argument,
libvmi.WithCloudInitNoCloudUserData has actually two functions. Let us
define these two functions with readable names.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>