Add the AlignCPUs enabler (#2695)
* api: Introduce the AlignCPUs enabler

PR kubevirt/kubevirt#10872 introduced a cluster-wide
setting to enable KubeVirt to request up to two additional
dedicated CPUs in order to complete the total CPU count
to an even parity when using emulator thread isolation.

Introduce a new enabler to HCO's API in order to toggle this
behavior in KubeVirt.

The new API field is optional and defaults to false.

The logic behind it will be introduced in the
following commits.

Signed-off-by: Orel Misan <omisan@redhat.com>
Signed-off-by: Ram Lavi <ralavi@redhat.com>

* Control KubeVirt AlignCPUs FG

Use the `AlignCPUs` enabler in order to control
KubeVirt's `AlignCPUs` feature gate.

Signed-off-by: Orel Misan <omisan@redhat.com>

* Introduce AlignCPUs logic

When the FG is enabled ensure the Kubevirt CR will have:
1. The `AlignCPUs` feature gate enabled.
2. The `alpha.kubevirt.io/EmulatorThreadCompleteToEvenParity`
annotation set with a value of an empty string.

Following PR kubevirt/kubevirt#10872, it will enable KubeVirt
to request up to two additional dedicated CPUs in order to complete the
total CPU count to an even parity when using emulator thread isolation.

When the FG is disabled, do the opposite:
1. Make sure the `AlignCPUs` feature gate is disabled.
2. Make sure the above annotation is not set.

The logic introduced in this commit is allowing users to freely
annotate the Kubevirt CR, but HCO will ensure the required
state of the above annotation.

Signed-off-by: Orel Misan <omisan@redhat.com>

* Docs: cluster-config: Document alignCPUs FG

Signed-off-by: Orel Misan <omisan@redhat.com>
Signed-off-by: Ram Lavi <ralavi@redhat.com>

---------

Signed-off-by: Orel Misan <omisan@redhat.com>
Signed-off-by: Ram Lavi <ralavi@redhat.com>