network bindings: Raise warning in case unregistered binding is used
In the scenario where an unregistered network binding plugin is being used
the VMI wont start until the plugin is registered.
To communicate this condition better and help with troubleshooting, raise
warning event in case an unregistered plugin is being used.

For example, given a VMI with the following interface spec:
spec:
  domain:
    devices:
      interfaces:
      - name: "mgmt"
        binding:
         name: "passt"
The following event will be raised:
$ kubectl describe vmi test-vmi
...
Events:
  Type     Reason                            Age                   From                       Message
  ----     ------                            ----                  ----                       -------
  Warning  UnregisteredNetworkBindingPlugin  44s (x17 over 6m11s)  virtualmachine-controller  network interface "mgmt" is set with unregistered network binding plugin "passt". Make sure all network binding plugins are registered in Kubevirt config.

This event emitting is rate limited as it being part of the reconcile loop
of the VMI controller, which is managed by the controller work queue.

Signed-off-by: Or Mergi <ormergi@redhat.com>