tests/operator: define usesSha as a static function
usesSha() is a single static function. Defining it as a named lambda
needlessly adds mental burden, as the read need to keep in mind that it
may change in the future.

With this change, there is no need to declare usesSha in one place, and
define it elsewhere, and indentation level is reduced.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>