fix: Reduce VM recovery times
It allows virt-controller to move a VMI to the `Failed` phase,
preventing cases where virt-handler is unavailable and VMI are stuck in
the `Running` phase.

In a scenario where a VM is running in a node which becomes unhealthy,
i.e. the node is no longer responsive. As a consequence, the launcher
pod is deleted, and the VM needs to be rescheduled if the spec
`runStrategy` is configured as `Always` or `RerunOnFailure`. In such
scenario, the VM's PHASE is `Running` but VM conditions is the
following:

```
Status:
  Conditions:
    Last Probe Time:       2023-12-18T16:31:31Z
    Last Transition Time:  2023-12-18T16:31:31Z
    Message:               virt-launcher pod has not yet been scheduled
    Reason:                PodNotExists
```

Eventually, the launcher pod is recreated. However, it takes a lot of
time.

The issue is: if a VMi in Phase `Running` on a given node, and the node
becomes unhealthy, VMis are not marked as `Failed`. This is due to the
fact that the main responsible for this task is the virt-handler pod,
which own the VMi. However, the virt-handler pod will be unresponsive,
unable to be redeployed and unable to mark the VMi as `Failed`.  The
secondary responsible for this action is the node controller.
Nevertheless, the node controller requires that heartbeats from the
virt-handler have not been received for a certain period of time (5
minutes).

Therefore, in order to transition the VMI to a `Failed` state, it is
required to wait until:

1. The node to get healthy again, the virt-handler to be recreated, the
   virt-handler to detect that VMi is on `Running` phase without a
launcher-pod, and to mark the VMi as `Failed`.

2. The node controller to detect that the virt-handler has been
   unresponsive for a given amount of time (5 minutes by default), to
detect VMis without virt-launcher pod and to mark the VMis as `Failed`.

It reduces the virt-launcher pod respawn time by short-circuiting the
feedback loop that the node controller would already handle and the
virt-handler recreation. Therefore, the virt-controller will mark the
VMi as `Failed` and it will force the VMi to be rescheduled faster.

NOTE: Slight change required with the original since `ptr` package is
not present in release branch 1.1.

Co-authored-by: Felix Matouschek <fmatouschek@redhat.com>
Signed-off-by: Javier Cano Cano <jcanocan@redhat.com>